Serena Dashboard time, shift, and y-axis changes
Changed the shift variable threshold so the chart will display 100
points instead of 20 before dropping the earlier points. Also, changed
the timeout variable to 1 second instead of 5 seconds. And finally, set
the min and max of the chart. The data being charted is implicitly
between 0 and 100; I just set the chart to those min and max at the
start.